+++ scanner/Options.py
@@ -127,9 +127,10 @@ config['browser_unsupported_mime_types'] = ['image/tiff', 'image/webp', 'image/x
 # json_version = 3.998 since symlinkCodesAndNumbers added as album property
 # json_version = 4.0 for new release
 # json_version = 4.1 since sizes of media, albums and trees added
+# json_version = 4.2 since numMediaInSubTree and numMedia are changed to numsMediaInSubTree and numsMedia
 
 # json_version = 0
-json_version = 4.1
+json_version = 4.2
 
 
 def set_obsolete_json_version_flag():