+++ web/js/036-functions.js
@@ -29,7 +29,7 @@
 		folders = location.pathname;
 		folders = folders.substring(0, folders.lastIndexOf('/'));
 		url += folders;
-		if (currentMedia === null || currentAlbum !== null && ! currentAlbum.subalbums.length && util.imagesAndVideosTotal(currentAlbum.numMedia) == 1) {
+		if (currentMedia === null || currentAlbum !== null && ! currentAlbum.subalbums.length && util.imagesAndVideosTotal(currentAlbum.numsMedia) == 1) {
 			mediaParameter = util.pathJoin([
 				Options.server_cache_path,
 				Options.cache_album_subdir,
@@ -515,8 +515,8 @@
 				currentMedia !== null ||
 				isAlbumWithOneMedia ||
 				thisAlbum !== null && (
-					util.imagesAndVideosTotal(thisAlbum.numMedia) === 0 ||
-					! util.isFolderCacheBase(thisAlbum.cacheBase) && util.imagesAndVideosTotal(thisAlbum.numMedia) > Options.big_virtual_folders_threshold
+					util.imagesAndVideosTotal(thisAlbum.numsMedia) === 0 ||
+					! util.isFolderCacheBase(thisAlbum.cacheBase) && util.imagesAndVideosTotal(thisAlbum.numsMedia) > Options.big_virtual_folders_threshold
 				)
 			)
 		) {
@@ -564,7 +564,7 @@
 
 		if (
 			thisAlbum === null ||
-			util.imagesAndVideosTotal(thisAlbum.numMedia) < Options.big_virtual_folders_threshold ||
+			util.imagesAndVideosTotal(thisAlbum.numsMedia) < Options.big_virtual_folders_threshold ||
 			util.isFolderCacheBase(thisAlbum.cacheBase)
 		) {
 			$("ul#right-menu #show-big-albums").addClass("hidden");
@@ -592,7 +592,7 @@
 				$("ul#right-menu li.album-sort").removeClass("hidden");
 			}
 
-			if (util.imagesAndVideosTotal(thisAlbum.numMedia) <= 1 || util.imagesAndVideosTotal(thisAlbum.numMedia) > Options.big_virtual_folders_threshold) {
+			if (util.imagesAndVideosTotal(thisAlbum.numsMedia) <= 1 || util.imagesAndVideosTotal(thisAlbum.numsMedia) > Options.big_virtual_folders_threshold) {
 				// no media or one media or too many media
 				$("ul#right-menu li.media-sort").addClass("hidden");
 			} else {
@@ -658,7 +658,7 @@
 			$(".select.everything").addClass("selected");
 		}
 
-		if (! Options.show_big_virtual_folders && util.imagesAndVideosTotal(thisAlbum.numMedia) > Options.big_virtual_folders_threshold) {
+		if (! Options.show_big_virtual_folders && util.imagesAndVideosTotal(thisAlbum.numsMedia) > Options.big_virtual_folders_threshold) {
 			$(".select.everything, .select.media").addClass("hidden");
 		}
 
@@ -669,7 +669,7 @@
 		if (
 			! thisAlbum.subalbums.length ||
 			! Options.show_big_virtual_folders && (
-				util.imagesAndVideosTotal(thisAlbum.numMediaInSubTree) > Options.big_virtual_folders_threshold
+				util.imagesAndVideosTotal(thisAlbum.numsMediaInSubTree) > Options.big_virtual_folders_threshold
 			)
 		) {
 			$(".select.everything-individual").addClass("hidden");
@@ -851,9 +851,9 @@
 				// reset the html
 				$(".download-album.everything.all").html(util._t(".download-album.everything.all"));
 
-				let nMediaInSubTree = util.imagesAndVideosTotal(thisAlbum.numMediaInSubTree);
-				let numImages = thisAlbum.numMediaInSubTree.images;
-				let numVideos = thisAlbum.numMediaInSubTree.videos;
+				let nMediaInSubTree = util.imagesAndVideosTotal(thisAlbum.numsMediaInSubTree);
+				let numImages = thisAlbum.numsMediaInSubTree.images;
+				let numVideos = thisAlbum.numsMediaInSubTree.videos;
 				let what = util._t(".title-media");
 				if (numImages === 0)
 					what = util._t(".title-videos");
@@ -861,9 +861,9 @@
 					what = util._t(".title-images");
 
 				if (util.isSearchCacheBase(thisAlbum.cacheBase) && thisAlbum.subalbums.length) {
-					// in search albums, numMediaInSubTree doesn't include the media in the albums found, the values that goes into the DOm must be update by code here
+					// in search albums, numsMediaInSubTree doesn't include the media in the albums found, the values that goes into the DOm must be update by code here
 					for (let iSubalbum = 0; iSubalbum < thisAlbum.subalbums.length; iSubalbum ++) {
-						nMediaInSubTree += util.imagesAndVideosTotal(thisAlbum.subalbums[iSubalbum].numMediaInSubTree);
+						nMediaInSubTree += util.imagesAndVideosTotal(thisAlbum.subalbums[iSubalbum].numsMediaInSubTree);
 					}
 				}
 
@@ -886,7 +886,7 @@
 						let reducedSize = Options.reduced_sizes[iSize];
 						treeSize = util.imagesAndVideosTotal(thisAlbum.sizesOfSubTree[reducedSize]);
 						if (treeSize < bigZipSize) {
-							$(".download-album.everything.all.sized").append(", " + reducedSize + " px: " + util.imagesAndVideosTotal(thisAlbum.numMediaInSubTree) + " " + what + ", " + Functions.humanFileSize(treeSize));
+							$(".download-album.everything.all.sized").append(", " + reducedSize + " px: " + util.imagesAndVideosTotal(thisAlbum.numsMediaInSubTree) + " " + what + ", " + Functions.humanFileSize(treeSize));
 							$(".download-album.everything.all.sized").attr("size", reducedSize);
 							$(".download-album.everything.all.sized").removeClass("hidden");
 							break;
@@ -905,8 +905,8 @@
 				// 	}
 				// }
 
-				let mediaInThisAlbum = util.imagesAndVideosTotal(thisAlbum.numMedia);
-				let mediaInThisTree = util.imagesAndVideosTotal(thisAlbum.numMediaInSubTree);
+				let mediaInThisAlbum = util.imagesAndVideosTotal(thisAlbum.numsMedia);
+				let mediaInThisTree = util.imagesAndVideosTotal(thisAlbum.numsMediaInSubTree);
 				if (numImages && numImages !== mediaInThisAlbum && numImages !== mediaInThisTree && mediaInThisAlbum !== mediaInThisTree) {
 					$(".download-album.everything.images.full").removeClass("hidden");
 					// reset the html
@@ -984,15 +984,15 @@
 			// 	}
 			// }
 			// TO DO: verify if it's correct to replace previous commented out code with the following 2 lines
-			let numImages = thisAlbum.numMedia.images;
-			let numVideos = thisAlbum.numMedia.videos;
+			let numImages = thisAlbum.numsMedia.images;
+			let numVideos = thisAlbum.numsMedia.videos;
 			let what = util._t(".title-media");
 			if (numImages === 0)
 				what = util._t(".title-videos");
 			if (numVideos === 0)
 				what = util._t(".title-images");
 
-			if (util.imagesAndVideosTotal(thisAlbum.numMedia)) {
+			if (util.imagesAndVideosTotal(thisAlbum.numsMedia)) {
 				$(".download-album.media-only.all.full").removeClass("hidden");
 				// reset the html
 				if (showDownloadEverything)
@@ -1002,7 +1002,7 @@
 
 				// add the download size
 				let albumSize = util.imagesAndVideosTotal(thisAlbum.sizesOfAlbum[0]);
-				$(".download-album.media-only.all.full").append(": " + util.imagesAndVideosTotal(thisAlbum.numMedia) + " " + what + ", " + Functions.humanFileSize(albumSize));
+				$(".download-album.media-only.all.full").append(": " + util.imagesAndVideosTotal(thisAlbum.numsMedia) + " " + what + ", " + Functions.humanFileSize(albumSize));
 				// check the size and decide if they can be downloaded
 				if (albumSize < bigZipSize) {
 					// maximum allowable size is 500MB (see https://github.com/eligrey/FileSaver.js/#supported-browsers)
@@ -1021,7 +1021,7 @@
 						let reducedSize = Options.reduced_sizes[iSize];
 						albumSize = thisAlbum.sizesOfAlbum[reducedSize].images + thisAlbum.sizesOfAlbum[reducedSize].videos;
 						if (albumSize < bigZipSize) {
-							$(".download-album.media-only.all.sized").append(", " + reducedSize + " px: " + util.imagesAndVideosTotal(thisAlbum.numMedia) + " " + what + ", " + Functions.humanFileSize(albumSize));
+							$(".download-album.media-only.all.sized").append(", " + reducedSize + " px: " + util.imagesAndVideosTotal(thisAlbum.numsMedia) + " " + what + ", " + Functions.humanFileSize(albumSize));
 							$(".download-album.media-only.all.sized").attr("size", reducedSize);
 							$(".download-album.media-only.all.sized").removeClass("hidden");
 							break;
@@ -1030,7 +1030,7 @@
 				}
 			}
 
-			if (numImages && numImages !== util.imagesAndVideosTotal(thisAlbum.numMedia)) {
+			if (numImages && numImages !== util.imagesAndVideosTotal(thisAlbum.numsMedia)) {
 				$(".download-album.media-only.images.full").removeClass("hidden");
 				// reset the html
 				if (showDownloadEverything)
@@ -1067,7 +1067,7 @@
 				}
 			}
 
-			if (numVideos && numVideos !== util.imagesAndVideosTotal(thisAlbum.numMedia)) {
+			if (numVideos && numVideos !== util.imagesAndVideosTotal(thisAlbum.numsMedia)) {
 				$(".download-album.media-only.videos.full").removeClass("hidden");
 				// reset the html
 				if (showDownloadEverything)