+++ doc/Changelog.md
@@ -1,3 +1,4 @@
+* numMediaInSubTree and numMedia album properties have been changed to numsMediaInSubTree and numsMedia (requieres json files regenerations)
 * new option `by_selection_string` for albums built from selection
 
 * menu refinements in download section